backend/services/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytube>=12.0.0
ffmpeg-python>=0.2.0
openai>=0.27.0
whisper>=1.0.0
TTS>=0.11.0
python-dotenv>=0.19.0